requirements.txt
fastapi[all]
slack-bolt==1.18.0
slack-sdk==3.21.3
aiohttp
google-cloud-firestore
google-cloud-logging
firebase-admin
pytest
gunicorn